requirements.txt
paho-mqtt
bluepy
pyyaml
schedule